+++ gdsfactory/types.py
@@ -43,7 +43,6 @@ Anchor = Literal[
     "center",
     "cc",
 ]
-SimulationSuffix = Literal[".dat", ".csv"]
 
 NSEW = Literal["N", "S", "E", "W"]